refactor: 로또 발급 service 인터페이스 제거(복잡도 해결차원)
