--- src/main/java/lotto/service/LottoIssuedService.java
@@ -1,8 +1,0 @@
-package lotto.service;
-
-import java.util.List;
-import java.util.Set;
-
-public interface LottoIssuedService {
-    Set<Integer> issueLottoNumber();
-}
