+++ src/test/java/lotto/domain/LottosTest.java
@@ -1,7 +1,5 @@
 package lotto.domain;
 
-import lotto.service.LottoAutoIssuedServiceImpl;
-import lotto.service.LottoIssuedService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -9,16 +7,16 @@ import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class LottosTest {
-    private LottoIssuedService lottoIssuedService;
+    private LottoMachine LottoMachine;
     @BeforeEach
     public void setUp() {
-        lottoIssuedService = new LottoAutoIssuedServiceImpl();
+        LottoMachine = new LottoMachine();
     }
 
     @Test
     @DisplayName("발급된 로또 갯수를 확인한다.")
     void 발급된_로또_갯수_확인() {
-        Lottos lottos = new Lottos(lottoIssuedService, 100);
+        Lottos lottos = new Lottos(LottoMachine, 100);
         assertThat(lottos.getLottos().size()).isEqualTo(100);
     }
 }
